pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quant_harbor"
version = "0.1.0"
description = "Local backtesting + dashboard"
requires-python = ">=3.9"

dependencies = [
  "backtrader",
  "pandas",
  "numpy",
  "pyarrow",
  "alpaca-py",
  "plotly",
  "streamlit",
  "pytz",
  "python-dateutil",
]
